move http error handling to http module
